requirements.txt
google-genai==1.25.0
openpyxl==3.1.5
pandas==2.3.1
pdfplumber==0.11.7
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1